Refactor network buffer types
